pyproject.toml
[tool.poetry]
name = "course-project"
version = "0.1.0"
description = ""
authors = ["LimDmitriy <160336426+LimDmitriy@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
poetry-core = "^2.1.3"
pandas = "^2.3.0"
python-dotenv = "^1.1.0"
openpyxl = "^3.1.5"
pandas-stubs = "^2.2.3.250527"
requests = "^2.32.4"


[tool.poetry.group.lint.dependencies]
mypy = "^1.16.0"
isort = "^6.0.1"
black = "^25.1.0"
flake8 = "^7.2.0"


[tool.poetry.group.dev.dependencies]
pytest = "^8.4.0"
pytest-cov = "^6.1.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.mypy]
disallow_untyped_defs = true
no_implicit_optional = true
warn_return_any = true
exclude = 'venv'

[tool.black]
# Максимальная длина строки
line-length = 119
# Файлы, которые не нужно форматировать
exclude = '''
(
  /(
      \.eggs         # Исключить несколько общих каталогов
    | \.git          # в корне проекта
    | \.hg
    | \.mypy_cache
    | \.tox
    | \.venv
    | dist
  )/
  | foo.py           # Также отдельно исключить файл с именем foo.py
                     # в корне проекта
)
'''

[tool.isort]
# максимальная длина строки
line_length = 119
